Code refactoring. Extracts booking calculation functions to a separate module. Imports them in components that require them.
